Kiosk: splitting out scanning logic
